ui/settings-window Move settings to standalone window
