+++ packages/electron-app/src/components/SettingsPage.tsx
@@ -1,575 +1,45 @@
-import { useState, useCallback, useEffect, useMemo } from 'react'
-import { ArrowLeft, Plus, Trash2, Layers } from 'lucide-react'
-import { Button } from '@/components/ui/button'
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog'
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select'
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
-import { useStore, useGlobalRoles } from '@/hooks/useStore'
-import { ROLE_COLOR_PALETTE } from '@/lib/constants'
-import { CreateRoleDialog } from '@/components/CreateRoleDialog'
-import { RoleFormFields } from '@/components/RoleFormFields'
-import { SetCollectionQuickAdd } from '@/components/SetCollectionQuickAdd'
-import { CacheSettingsSection } from '@/components/CacheSettingsSection'
-import { DataManagementSection } from '@/components/DataManagementSection'
-import { McpIntegrationCard } from '@/components/McpIntegrationCard'
+import { useState, useCallback } from 'react'
+import { SettingsSidebar, SETTINGS_SECTIONS, type SettingsSection } from '@/components/SettingsSidebar'
 import { ThemeSettingsSection } from '@/components/ThemeSettingsSection'
-import { getAllSets, type ScryfallSet } from '@/lib/scryfall'
-import type { RoleDefinition, SetCollectionEntry, CollectionLevel } from '@/types'
-import { getAllDeckEntries } from '@mtg-deckbuilder/shared'
+import { DataManagementSection } from '@/components/DataManagementSection'
+import { CacheSettingsSection } from '@/components/CacheSettingsSection'
+import { SetCollectionPane } from '@/components/settings/SetCollectionPane'
+import { RolesPane } from '@/components/settings/RolesPane'
+import { IntegrationsPane } from '@/components/settings/IntegrationsPane'
+
+function renderPane(section: SettingsSection) {
+  switch (section) {
+    case 'general':       return <ThemeSettingsSection />
+    case 'sets':          return <SetCollectionPane />
+    case 'roles':         return <RolesPane />
+    case 'integrations':  return <IntegrationsPane />
+    case 'data':          return <DataManagementSection />
+    case 'cache':         return <CacheSettingsSection />
+  }
+}
+
+function parseInitialSection(): SettingsSection {
+  const param = new URLSearchParams(window.location.search).get('section')
+  const match = SETTINGS_SECTIONS.find(s => s.id === param)
+  return match?.id ?? 'general'
+}
 
 export function SettingsPage() {
-  const setView = useStore(state => state.setView)
-  const globalRoles = useGlobalRoles()
-  const addGlobalRole = useStore(state => state.addGlobalRole)
-  const updateGlobalRole = useStore(state => state.updateGlobalRole)
-  const deleteGlobalRole = useStore(state => state.deleteGlobalRole)
-  const decks = useStore(state => state.decks)
-  const setCollection = useStore(state => state.setCollection)
-  const addSetToCollection = useStore(state => state.addSetToCollection)
-  const updateSetInCollection = useStore(state => state.updateSetInCollection)
-  const removeSetFromCollection = useStore(state => state.removeSetFromCollection)
+  const [active, setActive] = useState<SettingsSection>(parseInitialSection)
 
-  const [showAddDialog, setShowAddDialog] = useState(false)
-  const [setToDelete, setSetToDelete] = useState<SetCollectionEntry | null>(null)
-  const [showDeleteSetDialog, setShowDeleteSetDialog] = useState(false)
-  const [showEditDialog, setShowEditDialog] = useState(false)
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false)
-  const [editingRole, setEditingRole] = useState<RoleDefinition | null>(null)
-  const [roleToDelete, setRoleToDelete] = useState<RoleDefinition | null>(null)
-
-  // Edit form state
-  const [roleName, setRoleName] = useState('')
-  const [roleDescription, setRoleDescription] = useState('')
-  const [roleColor, setRoleColor] = useState<string>(ROLE_COLOR_PALETTE[0])
-
-  // Scryfall sets for looking up release years
-  const [allSets, setAllSets] = useState<ScryfallSet[]>([])
-
-  // Load sets for release year lookup
-  useEffect(() => {
-    getAllSets().then(setAllSets)
+  const handleChange = useCallback((section: SettingsSection) => {
+    setActive(section)
+    void window.electronAPI.saveSettingsActiveSection(section)
   }, [])
-
-  // Create lookup maps for release years and set types
-  const setInfo = useMemo(() => {
-    const releaseYears = new Map<string, string>()
-    const setTypes = new Map<string, string>()
-    for (const set of allSets) {
-      const code = set.code.toLowerCase()
-      if (set.released_at) {
-        releaseYears.set(code, set.released_at)
-      }
-      if (set.set_type) {
-        setTypes.set(code, set.set_type)
-      }
-    }
-    return { releaseYears, setTypes }
-  }, [allSets])
-
-  // Group labels for set types
-  const TYPE_GROUP_LABELS: Record<string, string> = {
-    'expansion': 'Expansions',
-    'core': 'Core Sets',
-    'draft_innovation': 'Draft Sets',
-    'masters': 'Masters Sets',
-    'commander': 'Commander',
-    'funny': 'Un-Sets',
-    'other': 'Other'
-  }
-
-  // Map set types to groups
-  const getTypeGroup = (setType: string): string => {
-    if (['expansion', 'core'].includes(setType)) return 'expansion'
-    if (setType === 'draft_innovation') return 'draft_innovation'
-    if (setType === 'masters') return 'masters'
-    if (setType === 'commander') return 'commander'
-    if (setType === 'funny') return 'funny'
-    return 'other'
-  }
-
-  const TYPE_GROUP_ORDER = ['expansion', 'draft_innovation', 'masters', 'commander', 'funny', 'other']
-
-  // Group and sort sets by type, then by release date within each group
-  const groupedSets = useMemo(() => {
-    if (!setCollection?.sets) return []
-
-    const groups = new Map<string, SetCollectionEntry[]>()
-
-    for (const entry of setCollection.sets) {
-      const setType = setInfo.setTypes.get(entry.setCode.toLowerCase()) || 'other'
-      const group = getTypeGroup(setType)
-      if (!groups.has(group)) {
-        groups.set(group, [])
-      }
-      groups.get(group)!.push(entry)
-    }
-
-    // Sort each group by release date (most recent first)
-    for (const [, entries] of groups) {
-      entries.sort((a, b) => {
-        const releasedA = a.releasedAt || setInfo.releaseYears.get(a.setCode.toLowerCase()) || ''
-        const releasedB = b.releasedAt || setInfo.releaseYears.get(b.setCode.toLowerCase()) || ''
-        return releasedB.localeCompare(releasedA)
-      })
-    }
-
-    // Build ordered result with group headers
-    const result: Array<{ type: 'header'; label: string } | { type: 'entry'; entry: SetCollectionEntry }> = []
-    for (const groupKey of TYPE_GROUP_ORDER) {
-      const entries = groups.get(groupKey)
-      if (entries && entries.length > 0) {
-        result.push({ type: 'header', label: TYPE_GROUP_LABELS[groupKey] || groupKey })
-        for (const entry of entries) {
-          result.push({ type: 'entry', entry })
-        }
-      }
-    }
-
-    return result
-  }, [setCollection?.sets, setInfo])
-
-  const resetForm = useCallback(() => {
-    setRoleName('')
-    setRoleDescription('')
-    setRoleColor(ROLE_COLOR_PALETTE[0])
-    setEditingRole(null)
-  }, [])
-
-  const handleAddRole = useCallback(async (role: RoleDefinition) => {
-    await addGlobalRole(role)
-  }, [addGlobalRole])
-
-  const handleEditRole = useCallback(async () => {
-    if (!editingRole || !roleName.trim()) return
-
-    await updateGlobalRole(editingRole.id, {
-      name: roleName.trim(),
-      description: roleDescription.trim() || undefined,
-      color: roleColor
-    })
-
-    resetForm()
-    setShowEditDialog(false)
-  }, [editingRole, roleName, roleDescription, roleColor, updateGlobalRole, resetForm])
-
-  const handleDeleteRole = useCallback(async () => {
-    if (!roleToDelete) return
-
-    await deleteGlobalRole(roleToDelete.id)
-    setRoleToDelete(null)
-    setShowDeleteDialog(false)
-  }, [roleToDelete, deleteGlobalRole])
-
-  const openEditDialog = useCallback((role: RoleDefinition) => {
-    setEditingRole(role)
-    setRoleName(role.name)
-    setRoleDescription(role.description || '')
-    setRoleColor(role.color || ROLE_COLOR_PALETTE[0])
-    setShowEditDialog(true)
-  }, [])
-
-  const openDeleteDialog = useCallback((role: RoleDefinition) => {
-    setRoleToDelete(role)
-    setShowDeleteDialog(true)
-  }, [])
-
-  // Set Collection handlers
-  const handleAddSet = useCallback(async (entry: Omit<SetCollectionEntry, 'addedAt'>) => {
-    await addSetToCollection(entry)
-  }, [addSetToCollection])
-
-  const handleDeleteSet = useCallback(async () => {
-    if (!setToDelete) return
-    await removeSetFromCollection(setToDelete.setCode)
-    setSetToDelete(null)
-    setShowDeleteSetDialog(false)
-  }, [setToDelete, removeSetFromCollection])
-
-  const handleInlineLevelChange = useCallback(async (setCode: string, level: CollectionLevel) => {
-    await updateSetInCollection(setCode, level)
-  }, [updateSetInCollection])
-
-  // Count how many cards/decks use a role
-  const getRoleUsage = useCallback((roleId: string) => {
-    let cardCount = 0
-    let deckCount = 0
-
-    decks.forEach(deck => {
-      const cardsWithRole = getAllDeckEntries(deck).filter(c => c.roles.includes(roleId))
-      if (cardsWithRole.length > 0) {
-        deckCount++
-        cardCount += cardsWithRole.length
-      }
-    })
-
-    return { cardCount, deckCount }
-  }, [decks])
 
   return (
-    <div className="h-full flex flex-col overflow-hidden">
-      {/* Header */}
-      <div className="p-6 pb-0 flex-shrink-0">
-        <div className="flex items-center gap-4 mb-6">
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => setView('decks')}
-            className="gap-2"
-          >
-            <ArrowLeft className="w-4 h-4" />
-            Back to Decks
-          </Button>
-          <h1 className="text-2xl font-bold">Settings</h1>
+    <div className="h-full flex overflow-hidden">
+      <SettingsSidebar active={active} onChange={handleChange} />
+      <main className="flex-1 overflow-auto">
+        <div className="max-w-3xl px-8 py-6">
+          {renderPane(active)}
         </div>
-      </div>
-
-      {/* Tabs */}
-      <Tabs defaultValue="collection" className="flex-1 flex flex-col overflow-hidden px-6">
-        <TabsList className="w-fit mb-6">
-          <TabsTrigger value="collection">Collection</TabsTrigger>
-          <TabsTrigger value="integration">Agent Integration</TabsTrigger>
-          <TabsTrigger value="system">System</TabsTrigger>
-        </TabsList>
-
-        {/* Collection Tab */}
-        <TabsContent value="collection" className="flex-1 overflow-auto m-0 pb-6">
-          <div className="max-w-3xl space-y-8">
-            {/* Set Collection Section */}
-            <section>
-              <h2 className="text-lg font-semibold mb-4">Set Collection</h2>
-
-              <p className="text-sm text-muted-foreground mb-4">
-                Track which MTG sets you own cards from. Collection levels determine which card rarities are included in Scryfall filters.
-              </p>
-
-              {/* Inline add bar */}
-              <div className="mb-4">
-                <SetCollectionQuickAdd
-                  onAdd={handleAddSet}
-                  existingSetCodes={setCollection?.sets.map(s => s.setCode.toLowerCase()) ?? []}
-                />
-              </div>
-
-              {/* Set table */}
-              {groupedSets.length > 0 ? (
-                <div className="border rounded-lg overflow-hidden max-h-80 overflow-y-auto">
-                  <table className="w-full text-sm table-fixed">
-                    <thead className="bg-muted sticky top-0">
-                      <tr className="text-left text-muted-foreground">
-                        <th className="px-3 py-2 font-medium w-14">Code</th>
-                        <th className="px-3 py-2 font-medium w-14">Year</th>
-                        <th className="px-3 py-2 font-medium">Set Name</th>
-                        <th className="px-3 py-2 font-medium w-28">Level</th>
-                        <th className="px-2 py-2 w-10"></th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {groupedSets.map((item) =>
-                        item.type === 'header' ? (
-                          <tr key={`header-${item.label}`} className="bg-muted/30">
-                            <td colSpan={5} className="px-3 py-1.5 text-xs font-semibold text-muted-foreground uppercase tracking-wide">
-                              {item.label}
-                            </td>
-                          </tr>
-                        ) : (
-                          <tr key={item.entry.setCode} className="hover:bg-muted/20 border-t border-border/50">
-                            <td className="px-3 py-1.5 uppercase text-muted-foreground font-mono text-xs">
-                              {item.entry.setCode}
-                            </td>
-                            <td className="px-3 py-1.5 text-muted-foreground">
-                              {(() => {
-                                const released = item.entry.releasedAt || setInfo.releaseYears.get(item.entry.setCode.toLowerCase())
-                                return released ? new Date(released).getFullYear() : '—'
-                              })()}
-                            </td>
-                            <td className="px-3 py-1.5 truncate" title={item.entry.setName}>
-                              {item.entry.setName}
-                            </td>
-                            <td className="px-3 py-1.5">
-                              <Select
-                                value={String(item.entry.collectionLevel)}
-                                onValueChange={(value) => handleInlineLevelChange(item.entry.setCode, Number(value) as CollectionLevel)}
-                              >
-                                <SelectTrigger className="h-7 w-24 text-xs">
-                                  <SelectValue />
-                                </SelectTrigger>
-                                <SelectContent>
-                                  {([1, 2, 3, 4] as CollectionLevel[]).map((level) => (
-                                    <SelectItem key={level} value={String(level)}>
-                                      Level {level}
-                                    </SelectItem>
-                                  ))}
-                                </SelectContent>
-                              </Select>
-                            </td>
-                            <td className="px-2 py-1.5">
-                              <Button
-                                variant="ghost"
-                                size="icon"
-                                className="h-6 w-6 text-muted-foreground hover:text-destructive"
-                                onClick={() => {
-                                  setSetToDelete(item.entry)
-                                  setShowDeleteSetDialog(true)
-                                }}
-                              >
-                                <Trash2 className="w-3.5 h-3.5" />
-                              </Button>
-                            </td>
-                          </tr>
-                        )
-                      )}
-                    </tbody>
-                  </table>
-                </div>
-              ) : (
-                <div className="text-center py-6 text-sm text-muted-foreground border rounded-lg">
-                  No sets in your collection yet. Search above to add sets.
-                </div>
-              )}
-            </section>
-
-            {/* Global Roles Section */}
-            <section>
-              <div className="flex items-center justify-between mb-4">
-                <h2 className="text-lg font-semibold">Global Roles</h2>
-                <Button
-                  onClick={() => {
-                    resetForm()
-                    setShowAddDialog(true)
-                  }}
-                  size="sm"
-                  className="gap-2"
-                >
-                  <Plus className="w-4 h-4" />
-                  Add Role
-                </Button>
-              </div>
-
-              <p className="text-sm text-muted-foreground mb-4">
-                Global roles are available across all decks. Hover for details, click to edit.
-              </p>
-
-              <TooltipProvider delayDuration={300}>
-                <div className="flex flex-wrap gap-2">
-                  {globalRoles.map(role => {
-                    const { cardCount, deckCount } = getRoleUsage(role.id)
-                    return (
-                      <Tooltip key={role.id}>
-                        <TooltipTrigger asChild>
-                          <button
-                            onClick={() => openEditDialog(role)}
-                            className="group flex items-center gap-1.5 px-2.5 py-1 rounded-full border bg-card hover:bg-accent transition-colors text-sm"
-                          >
-                            <div
-                              className="w-2.5 h-2.5 rounded-full flex-shrink-0"
-                              style={{ backgroundColor: role.color || '#888' }}
-                            />
-                            <span className="font-medium">{role.name}</span>
-                            {deckCount > 0 && (
-                              <span className="flex items-center gap-0.5 text-xs text-muted-foreground">
-                                <Layers className="w-3 h-3" />
-                                {deckCount}
-                              </span>
-                            )}
-                            <Trash2
-                              className="w-3 h-3 text-muted-foreground opacity-0 group-hover:opacity-100 hover:text-destructive transition-opacity ml-0.5"
-                              onClick={(e) => {
-                                e.stopPropagation()
-                                openDeleteDialog(role)
-                              }}
-                            />
-                          </button>
-                        </TooltipTrigger>
-                        <TooltipContent side="bottom" className="max-w-xs">
-                          <div className="space-y-1">
-                            {role.description && (
-                              <p className="text-sm">{role.description}</p>
-                            )}
-                            {cardCount > 0 ? (
-                              <p className="text-xs text-muted-foreground">
-                                Used by {cardCount} card{cardCount !== 1 ? 's' : ''} in {deckCount} deck{deckCount !== 1 ? 's' : ''}
-                              </p>
-                            ) : (
-                              <p className="text-xs text-muted-foreground">Not used yet</p>
-                            )}
-                          </div>
-                        </TooltipContent>
-                      </Tooltip>
-                    )
-                  })}
-                </div>
-              </TooltipProvider>
-
-              {globalRoles.length === 0 && (
-                <div className="text-center py-8 text-muted-foreground">
-                  No global roles defined. Add some roles to get started.
-                </div>
-              )}
-            </section>
-          </div>
-        </TabsContent>
-
-        {/* Agent Integration Tab */}
-        <TabsContent value="integration" className="flex-1 overflow-auto m-0 pb-6">
-          <div className="max-w-3xl space-y-6">
-            <McpIntegrationCard
-              clientId="claude-desktop"
-              title="Claude Desktop"
-              connectedDescription="Claude can help you manage your decks through conversation"
-              disconnectedDescription="Connect to use AI-powered deck building features"
-              connectButtonLabel="Connect to Claude"
-              postConnectionNote="You may need to restart Claude Desktop for changes to take effect."
-            />
-            <McpIntegrationCard
-              clientId="claude-code"
-              title="Claude Code"
-              connectedDescription="Claude Code can access your deck data through MCP tools"
-              disconnectedDescription="Connect to enable deck management from Claude Code"
-              connectButtonLabel="Connect to Claude Code"
-              postConnectionNote="You may need to restart Claude Code or re-enter your project for changes to take effect."
-            />
-            <McpIntegrationCard
-              clientId="gemini-cli"
-              title="Gemini CLI"
-              connectedDescription="Gemini can help you manage your decks through the CLI"
-              disconnectedDescription="Connect to use Gemini CLI for deck building"
-              connectButtonLabel="Connect to Gemini CLI"
-              postConnectionNote="You may need to restart Gemini CLI for changes to take effect."
-            />
-          </div>
-        </TabsContent>
-
-        {/* System Tab */}
-        <TabsContent value="system" className="flex-1 overflow-auto m-0 pb-6">
-          <div className="max-w-3xl space-y-8">
-            <section>
-              <h2 className="text-lg font-semibold mb-4">Appearance</h2>
-              <ThemeSettingsSection />
-            </section>
-            <section>
-              <h2 className="text-lg font-semibold mb-4">Data Management</h2>
-              <DataManagementSection />
-            </section>
-            <section>
-              <h2 className="text-lg font-semibold mb-4">Scryfall Cache</h2>
-              <CacheSettingsSection />
-            </section>
-          </div>
-        </TabsContent>
-      </Tabs>
-
-      {/* Add Role Dialog */}
-      <CreateRoleDialog
-        open={showAddDialog}
-        onOpenChange={setShowAddDialog}
-        onCreated={handleAddRole}
-        title="Add Global Role"
-        description="Create a new role that will be available across all decks."
-      />
-
-      {/* Edit Role Dialog */}
-      <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Edit Role</DialogTitle>
-            <DialogDescription>
-              Update the role definition.
-            </DialogDescription>
-          </DialogHeader>
-          <div className="py-4">
-            <RoleFormFields
-              roleName={roleName}
-              onRoleNameChange={setRoleName}
-              roleDescription={roleDescription}
-              onRoleDescriptionChange={setRoleDescription}
-              roleColor={roleColor}
-              onRoleColorChange={setRoleColor}
-              onSubmit={handleEditRole}
-            />
-          </div>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setShowEditDialog(false)}>
-              Cancel
-            </Button>
-            <Button onClick={handleEditRole} disabled={!roleName.trim()}>
-              Save Changes
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      {/* Delete Confirmation Dialog */}
-      <Dialog open={showDeleteDialog} onOpenChange={setShowDeleteDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Delete Role</DialogTitle>
-            <DialogDescription>
-              {roleToDelete && (() => {
-                const { cardCount, deckCount } = getRoleUsage(roleToDelete.id)
-                if (cardCount > 0) {
-                  return `This role is used by ${cardCount} card${cardCount !== 1 ? 's' : ''} across ${deckCount} deck${deckCount !== 1 ? 's' : ''}. Are you sure you want to delete "${roleToDelete.name}"? Cards will keep the role ID but it won't appear in the role list.`
-                }
-                return `Are you sure you want to delete "${roleToDelete.name}"? This action cannot be undone.`
-              })()}
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setShowDeleteDialog(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={handleDeleteRole}
-            >
-              Delete
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      {/* Delete Set Confirmation Dialog */}
-      <Dialog open={showDeleteSetDialog} onOpenChange={setShowDeleteSetDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Remove Set</DialogTitle>
-            <DialogDescription>
-              {setToDelete && (
-                <>Are you sure you want to remove "{setToDelete.setName}" from your collection?</>
-              )}
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setShowDeleteSetDialog(false)}>
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              onClick={handleDeleteSet}
-            >
-              Remove
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      </main>
     </div>
   )
 }
