+++ packages/electron-app/src/main.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { App } from './App'
+import { SettingsWindow } from './SettingsWindow'
 import { ErrorBoundary } from './components/ErrorBoundary'
 import './styles/fonts'
 import './styles/globals.css'
@@ -15,11 +16,13 @@ const queryClient = new QueryClient({
   },
 })
 
+const isSettingsWindow = new URLSearchParams(window.location.search).get('view') === 'settings'
+
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <ErrorBoundary>
-        <App />
+        {isSettingsWindow ? <SettingsWindow /> : <App />}
       </ErrorBoundary>
     </QueryClientProvider>
   </React.StrictMode>
