+++ packages/electron-app/src/stores/types.ts
@@ -1,7 +1,7 @@
 import type { Deck, Taxonomy, CardList, Config, CardEntry, DeckNote, CardIdentifier, RoleDefinition, SetCollectionFile, SetCollectionEntry, CollectionLevel, PullListConfig, CardSetName, CardSource, ArtCardFace } from '@/types'
 import type { UISlice } from './uiSlice'
 
-export type AppView = 'decks' | 'deck-detail' | 'interest-list' | 'buy-list' | 'settings'
+export type AppView = 'decks' | 'deck-detail' | 'interest-list' | 'buy-list'
 
 export interface AppState
   extends DeckSlice,
